добавлено логирование в репозиторий
